+++ src/ViaAuthorization/Resolver.php
@@ -5,12 +5,12 @@ namespace Innmind\HttpAuthentication\ViaAuthorization;
 
 use Innmind\HttpAuthentication\Identity;
 use Innmind\Http\Header\AuthorizationValue;
-use Innmind\Immutable\Maybe;
+use Innmind\Immutable\Attempt;
 
 interface Resolver
 {
     /**
-     * @return Maybe<Identity>
+     * @return Attempt<Identity>
      */
-    public function __invoke(AuthorizationValue $value): Maybe;
+    public function __invoke(AuthorizationValue $value): Attempt;
 }