+++ src/ViaBasicAuthorization/NullResolver.php
@@ -4,13 +4,13 @@ declare(strict_types = 1);
 namespace Innmind\HttpAuthentication\ViaBasicAuthorization;
 
 use Innmind\HttpAuthentication\Identity;
-use Innmind\Immutable\Maybe;
+use Innmind\Immutable\Attempt;
 
 final class NullResolver implements Resolver
 {
-    public function __invoke(string $user, string $password): Maybe
+    public function __invoke(string $user, string $password): Attempt
     {
-        /** @var Maybe<Identity> */
-        return Maybe::nothing();
+        /** @var Attempt<Identity> */
+        return Attempt::error(new \LogicException('Not implemented'));
     }
 }