+++ src/ViaUrlAuthority/NullResolver.php
@@ -8,13 +8,13 @@ use Innmind\Url\Authority\UserInformation\{
     User,
     Password,
 };
-use Innmind\Immutable\Maybe;
+use Innmind\Immutable\Attempt;
 
 final class NullResolver implements Resolver
 {
-    public function __invoke(User $user, Password $password): Maybe
+    public function __invoke(User $user, Password $password): Attempt
     {
-        /** @var Maybe<Identity> */
-        return Maybe::nothing();
+        /** @var Attempt<Identity> */
+        return Attempt::error(new \LogicException('Not implemented'));
     }
 }