+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - `Innmind\HttpAuthentication\Authenticator::__invoke()` now returns an `Innmind\Immutable\Attempt`
+- All resolvers now return an `Innmind\Immutable\Attempt`
 
 ## 4.0.0 - 2023-11-01